Refactor time handling and data types in coinflip program
- Updated the `current_time` parameter type in several methods from `u64` to `i64` to improve consistency in time representation.
- Adjusted calculations in the `ready_for_oracle`, `buffer_passed`, and `calculate_winner` methods to accommodate the new `i64` type, ensuring correct logic for game readiness and winner determination.
- Modified the `SetOracleHash`, `UnjoinGame`, and `CancelGame` instructions to reflect the updated time handling, enhancing the integrity of game state validation.

These changes enhance the maintainability and clarity of the coinflip program by refining time handling and ensuring consistent data types across the codebase.

+++ programs/coinflip/src/instructions/accounts.rs
@@ -278,7 +278,7 @@ pub struct SetOracleHash<'info> {
         constraint = game_token.token_mint == game.token_mint @ ErrorCode::InvalidToken,
     )]
     #[account(
-        seeds = [b"player_token", game.calculate_winner(random_number as usize, Clock::get()?.unix_timestamp as u64).as_ref(), game_token.token_mint.as_ref()],
+        seeds = [b"player_token", game.calculate_winner(random_number, Clock::get()?.unix_timestamp).as_ref(), game_token.token_mint.as_ref()],
         bump,
     )]
     pub player_token: Account<'info, PlayerToken>,
@@ -293,7 +293,7 @@ pub struct UnjoinGame<'info> {
         mut,
         constraint = game.status == GameStatus::Active @ ErrorCode::GameNotActive,
         constraint = game.creator != player.key() && game.players.contains(&player.key()) @ ErrorCode::UnauthorizedPlayer,
-        constraint = !game.ready_for_oracle(Clock::get()?.unix_timestamp as u64) @ ErrorCode::GameReadyForOracle,
+        constraint = !game.ready_for_oracle(Clock::get()?.unix_timestamp) @ ErrorCode::GameReadyForOracle,
     )]
     pub game: Account<'info, Game>,
     pub player: Signer<'info>,
@@ -335,7 +335,7 @@ pub struct CancelGame<'info> {
         mut,
         constraint = game.status == GameStatus::Active @ ErrorCode::GameNotActive,
         constraint = game.creator == player.key() || (game.players.contains(&player.key()) && game.game_type != GameType::Giveaway) @ ErrorCode::UnauthorizedPlayer,
-        constraint = game.buffer_passed(oracle.oracle_buffer_time, Clock::get()?.unix_timestamp as u64) @ ErrorCode::GameReadyForOracle,
+        constraint = game.buffer_passed(oracle.oracle_buffer_time, Clock::get()?.unix_timestamp) @ ErrorCode::GameReadyForOracle,
     )]
     pub game: Account<'info, Game>,
     #[account()]
